only deleting GPU kernel if it's available (for cases when going through the GPU code path on CPU-only hardware)

+++ cppn-neat/network-activation.js
@@ -563,7 +563,7 @@ class Activator {
             try {
               // console.log("---oneOutputKernel:",oneOutputKernel);
               let outputResult = oneOutputKernel();
-              oneOutputKernel.kernel.texture.delete();
+              if(oneOutputKernel.kernel && oneOutputKernel.kernel.texture) oneOutputKernel.kernel.texture.delete();
               resolve( outputResult );
             } catch (e) {
               reject( e );
